change to use gint instead of gint16, to avoid overflow issues, since
2001-03-16  Havoc Pennington  <hp@redhat.com>

	* gtk/gtkwidget.h (struct _GtkWidgetAuxInfo): change to use gint
	instead of gint16, to avoid overflow issues, since sizes > 16-bit
	are now allowed.

2001-03-16  Havoc Pennington  <hp@redhat.com>

	* gtk/gtk-docs.sgml: remove GtkData entity usage

+++ gtk/gtkwidget.h
@@ -380,10 +380,10 @@ struct _GtkWidgetClass
 
 struct _GtkWidgetAuxInfo
 {
-  gint16  x;
-  gint16  y;
-  gint16 width;
-  gint16 height;
+  gint x;
+  gint y;
+  gint width;
+  gint height;
 };
 
 struct _GtkWidgetShapeInfo
